Nest the moved controller under pkg/enterprise/controller
The Enterprise tree mirrors the core tree it came from, keeping controllers
apart from the extension packages that share the directory.
