+++ pkg/enterprise/csr/extension_test.go
@@ -39,7 +39,7 @@ import (
 	"github.com/tigera/operator/pkg/ctrlruntime"
 	ctrlrfake "github.com/tigera/operator/pkg/ctrlruntime/client/fake"
 	"github.com/tigera/operator/pkg/dns"
-	"github.com/tigera/operator/pkg/enterprise/monitor"
+	"github.com/tigera/operator/pkg/enterprise/controller/monitor"
 	"github.com/tigera/operator/pkg/render"
 	rmonitor "github.com/tigera/operator/pkg/render/monitor"
 )
